refactored existing tests with ast, added 2 new that ensure #19 is solved
